pyproject.toml
[project]
name = "poe-wrapper"
version = "0.1.0"
description = "Асинхронная прослойка для Poe REST API"
readme = "README.md"
authors = [
    { name = "femto" }
]
requires-python = ">=3.13.11"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi-poe>=0.0.83",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
